Provide CMake shims to ease working with UPS
The cetbuildtools/UPS build workflow requires quite a bit of care
and knowledge to run quickly and correctly. This is mostly an artifact
of heavy back-and-forth between ups/product_deps,
ups/setup_for_development, and CMake.

As we only require a very simple build and want as familiar an
environment as possible, provide CMake shims pre- and post-project()
call to implement behaviour of "setup_for_development":

- Set CC/CXX to required UPS compiler names "gcc/g++" rather than
  the "cc/c++" that CMake will find by default
- Set default build mode based on that art of qualifier
- Generate the expect "cetpkg_variable_report" file expected by UPS

Update documents to show how to build using this method, even though
nothing is compiled just yet.

NB: a ups/product_deps file is still required for now. Needs further
checks to confirm how this info is used in cetbuildtools.