Demo FHiCL ParameterSetID behaviour
FHiCL's ParameterSet objects can be hashed to give a unique id based
on a hash of the key/value contents. This can be useful for data
cataloguing, so demo behaviour with a simple application to check
the hashes of two ParameterSets made from input files. Test that
the ParameterSetIDs for the reference set is identical to that
of other parameter sets that should have the same content albeit
constructed in different ways.